requirements.txt
python-telegram-bot==21.1
apscheduler==3.10.4
aiofiles==23.2.1
python-dotenv==1.0.0
supabase==2.28.3
websockets==15.0.1
